fixup! chore(actionpack): wave 3 — AbstractController split into abstractcontroller/ top-level dir
- Remove backward-compat stub (actioncontroller/abstract-controller.ts)
- Update all internal imports to canonical abstractcontroller/ paths
- Fix error.ts: export AbstractControllerError (AbstractController::Error)
  per Rails error.rb; ActionNotFound moves to base.ts where Rails defines it
- Move abstract-controller.test.ts → abstractcontroller/callbacks.test.ts
  to match Rails test/abstract/callbacks_test.rb location
