--- packages/actionpack/src/actioncontroller/abstract-controller.ts
@@ -1,7 +1,0 @@
-export {
-  AbstractController,
-  ActionNotFound,
-  type ActionCallback,
-  type AroundCallback,
-  type CallbackOptions,
-} from "../abstractcontroller/base.js";
